+++ examples/src/tes/main.rs
@@ -23,11 +23,15 @@ use crankshaft::engine::Task;
 use crankshaft::engine::task::Execution;
 use eyre::Result;
 use eyre::bail;
+use futures::FutureExt;
 use futures::StreamExt;
 use futures::stream::FuturesUnordered;
 use indicatif::ProgressBar;
 use indicatif::ProgressStyle;
 use nonempty::NonEmpty;
+use tokio::select;
+use tokio::signal;
+use tokio_util::sync::CancellationToken;
 use tracing_subscriber::EnvFilter;
 use tracing_subscriber::fmt;
 use tracing_subscriber::layer::SubscriberExt as _;
@@ -56,7 +60,7 @@ pub struct Args {
 }
 
 /// Starting point for task execution.
-async fn run(args: Args) -> Result<()> {
+async fn run(args: Args, token: CancellationToken) -> Result<()> {
     let config = Config::builder().url(args.url);
 
     let username = std::env::var(USER_ENV).ok();
@@ -104,7 +108,7 @@ async fn run(args: Args) -> Result<()> {
     Engine::start_instrument(3000);
 
     let mut tasks = (0..args.n_jobs)
-        .map(|_| Ok(engine.spawn("tes", task.clone())?.wait()))
+        .map(|_| Ok(engine.spawn("tes", task.clone(), token.clone())?.wait()))
         .collect::<Result<FuturesUnordered<_>>>()?;
 
     let progress = ProgressBar::new(tasks.len() as u64);
@@ -150,7 +154,9 @@ async fn run(args: Args) -> Result<()> {
     Ok(())
 }
 
-fn main() -> Result<()> {
+/// The main function.
+#[tokio::main]
+async fn main() -> Result<()> {
     let args = Args::parse();
 
     tracing_subscriber::registry()
@@ -158,9 +164,15 @@ fn main() -> Result<()> {
         .with(EnvFilter::from_default_env())
         .init();
 
-    tokio::runtime::Builder::new_multi_thread()
-        .enable_all()
-        .build()
-        .unwrap()
-        .block_on(run(args))
+    let cancellation = CancellationToken::new();
+    let mut run = run(args, cancellation.clone()).boxed();
+
+    select! {
+        _ = signal::ctrl_c() => {
+            eprintln!("\nexecution was interrupted; waiting for tasks to cancel");
+            cancellation.cancel();
+            run.await
+        },
+        res = &mut run => return res,
+    }
 }